+++ src/Exception/QiqModuleException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BEAR\QiqModule\Exception;
+
+use RuntimeException;
+
+abstract class QiqModuleException extends RuntimeException
+{
+}