+++ src/QiqProdModule.php
@@ -4,22 +4,22 @@ declare(strict_types=1);
 
 namespace BEAR\QiqModule;
 
+use BEAR\Sunday\Compile\CompileStepInterface;
+use Qiq\Catalog;
 use Qiq\Compiler;
-use Qiq\Compiler\QiqCompiler;
+use Qiq\Compiler\NonCompiler;
 use Ray\Di\AbstractModule;
+use Ray\Di\MultiBinder;
 use ReflectionException;
 
 final class QiqProdModule extends AbstractModule
 {
-    public function __construct(private string $cachePath, private AbstractModule|null $module = null)
-    {
-        parent::__construct($this->module);
-    }
-
     /** @throws ReflectionException */
     protected function configure(): void
     {
-        $this->bind()->annotatedWith('qiq_cache_path')->toInstance($this->cachePath);
-        $this->bind(Compiler::class)->toConstructor(QiqCompiler::class, ['cachePath' => 'qiq_cache_path']);
+        MultiBinder::newInstance($this, CompileStepInterface::class)
+            ->addBinding(QiqCompileStep::NAME)->to(QiqCompileStep::class);
+        $this->bind(Catalog::class)->to(QiqProdCatalog::class);
+        $this->bind(Compiler::class)->to(NonCompiler::class);
     }
 }